split the h files from the cc files and added a makefile
