+++ .gitignore
@@ -1,0 +1,5 @@
+*~
+*.o
+client/client
+client/graphics_test
+client/test
