--- client/flyingObject.cc~
@@ -1,19 +1,0 @@
-class flyingObject{
-  float* position;
-  float* velocity;
-  float* acceleration;
-  float* orientation;
-  
-public:
-  flyingObject();
-  ~flyingObject();
-}
-
-
-void flyingObject::flyingObject(){
-  
-}
-
-void flyingObject::~flyingObject(){
-  
-}
