--- client/graphics_test.cc~
@@ -1,53 +1,0 @@
-#include <GL/glut.h>
-#include <stdio.h>
-#include <stdlib.h>
-#define USING_MATH_DEFINES
-#include <math.h>
-#include <float.h>
-#include <assert.h>
-
-#define THETA_STEP 2 //degrees
-#define PHI_STEP   2 //degrees
-
-void display(void);
-void keypress(unsigned char key, int x, int y);
-
-int main(int argc, char **argv){
-  
-  glutInit(&argc, argv);
-  glutInitDisplayMode(GLUT_DOUBLE | GLUT_RGB | GLUT_DEPTH);
-  glutCreateWindow("graphics_test");
-  glutFullScreen();
-  
-  glutDisplayFunc(display);
-  glutKeyboardFunc(keypress);
-  
-  
-  
-  glutMainLoop();
-  
-  return 0;
-}
-
-
-void keypress(unsigned char key, int x, int y){
-  switch (key) {
-  case 27:
-    exit(0);
-    break;
-  }
-}
-
-void display(void){
-  int i, j;
-  float dist = 0;
-
-  //Clear screen
-  glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-  glEnd();
-  
-  glPopMatrix();
-
-  glutSwapBuffers();
-}
