--- client/joy_test.cc~
@@ -1,7 +1,0 @@
-#include "Joystick.h"
-
-int main(int argc, char** argv){
-  Joystick joystick;
-  
-  return 0;
-}
