--- client/test.cc~
@@ -1,8 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-
-#include "flyingObject.cc"
-
-int main(int argc, char **argv){
-  return 0;
-}
